Extract apps installation code to useActiveAppsInstallations hook
